// src/modules/types.ts
export interface User {
  firstName: string
  email: string
  id: string
  createdAt: any
}

export interface AuthState {
  user: User | null
  authenticated: boolean
  loading: boolean
  error: string
  needVerification: boolean
  success: string
}

export interface SignUpData {
  firstName: string
  email: string
  password: string
}

export interface SignInData {
  email: string
  password: string
}